+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 - fix security bundle salt configuration for itsdangerous 2.0+
 - add shell readline completion (from [Flask PR 3960](https://github.com/pallets/flask/pull/3960))
-- fix `BundleBlueprint.register` to work with Flask 2.0+ 
+- fix `BundleBlueprint.register` to work with Flask 2.0+
+- bump required flask-sqlalchemy-unchained version
 
 ## v0.8.1 (2021/01/17)