+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - add shell readline completion (from [Flask PR 3960](https://github.com/pallets/flask/pull/3960))
 - fix `BundleBlueprint.register` to work with Flask 2.0+
 - bump required flask-sqlalchemy-unchained version
+- API bundle fixes and improvements
 
 ## v0.8.1 (2021/01/17)